Update.
	* sysdeps/unix/sysv/linux/bits/socket.h (CMSG_ALIGN): Cast result
	of unary ~ to size_t.

	* stdlib/strtod.c: Make gcc shut up about SWAP use.

+++ sysdeps/unix/sysv/linux/bits/socket.h
@@ -244,7 +244,7 @@ struct cmsghdr
   ((size_t) (mhdr)->msg_controllen >= sizeof (struct cmsghdr)		      \
    ? (struct cmsghdr *) (mhdr)->msg_control : (struct cmsghdr *) NULL)
 #define CMSG_ALIGN(len) (((len) + sizeof (size_t) - 1) \
-			 & ~(sizeof (size_t) - 1))
+			 & (size_t) ~(sizeof (size_t) - 1))
 #define CMSG_SPACE(len) (CMSG_ALIGN (len) \
 			 + CMSG_ALIGN (sizeof (struct cmsghdr)))
 #define CMSG_LEN(len)   (CMSG_ALIGN (sizeof (struct cmsghdr)) + (len))
